// src/app/model/items.ts
export class Item{
    candidateId: number= 0;
    candidateName: string= " ";
    candidateCategory: string= " ";
    candidateEmail: string= " ";
    candidatePhone: number= 0;
    candidateGender: string= " ";
    candidateAddress: string= " ";            
    candidatePincode: number= 0;

    constructor(){
        
    }
}